+++ contracts/token/ERC20/ERC20.sol
@@ -30,11 +30,12 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 pragma solidity ^0.6.8;
 
 import "@openzeppelin/contracts/GSN/Context.sol";
-import "@openzeppelin/contracts/introspection/IERC165.sol";
+import "@openzeppelin/contracts/introspection/ERC165.sol";
 import "@openzeppelin/contracts/math/SafeMath.sol";
 import "@openzeppelin/contracts/utils/Address.sol";
 import "./IERC20.sol";
 import "./IERC20Detailed.sol";
+import "./IERC20Allowance.sol";
 
 /**
  * @dev Implementation of the {IERC20} interface.
@@ -60,7 +61,7 @@ import "./IERC20Detailed.sol";
  * functions have been added to mitigate the well-known issues around setting
  * allowances. See {IERC20-approve}.
  */
-abstract contract ERC20 is Context, IERC165, IERC20, IERC20Detailed {
+abstract contract ERC20 is ERC165, Context, IERC20, IERC20Detailed, IERC20Allowance {
 
     using SafeMath for uint256;
     using Address for address;
@@ -71,27 +72,20 @@ abstract contract ERC20 is Context, IERC165, IERC20, IERC20Detailed {
 
     uint256 internal _totalSupply;
 
-    /**
-     * @dev Check if support an interface id
-     * @param interfaceId interface id to query
-     * @return bool if support the given interface id
-     */
-    function supportsInterface(bytes4 interfaceId) external virtual override view returns (bool) {
-        return (
-            // ERC165 interface id
-            interfaceId == 0x01ffc9a7 ||
-            // ERC20 interface id
-            interfaceId == 0x36372b07 ||
-            // ERC20Name interface id
-            interfaceId == 0x06fdde03 ||
-            // ERC20Symbol interface id
-            interfaceId == 0x95d89b41 ||
-            // ERC20Decimals interface id
-            interfaceId == 0x313ce567 ||
-            // ERC20Detailed interface id
-            interfaceId == 0xa219a025
-        );
+    constructor() internal {
+        _registerInterface(type(IERC20).interfaceId);
+        _registerInterface(type(IERC20Detailed).interfaceId);
+        _registerInterface(type(IERC20Allowance).interfaceId);
+
+        // ERC20Name interfaceId: bytes4(keccak256("name()"))
+        _registerInterface(0x06fdde03);
+        // ERC20Symbol interfaceId: bytes4(keccak256("symbol()"))
+        _registerInterface(0x95d89b41);
+        // ERC20Decimals interfaceId: bytes4(keccak256("decimals()"))
+        _registerInterface(0x313ce567);
     }
+
+/////////////////////////////////////////// ERC20 ///////////////////////////////////////
 
     /**
      * @dev See {IERC20-totalSupply}.
@@ -157,41 +151,33 @@ abstract contract ERC20 is Context, IERC165, IERC20, IERC20Detailed {
         return true;
     }
 
+/////////////////////////////////////////// ERC20Allowance ///////////////////////////////////////
+
     /**
-     * @dev Atomically increases the allowance granted to `spender` by the caller.
-     *
-     * This is an alternative to {approve} that can be used as a mitigation for
-     * problems described in {IERC20-approve}.
-     *
-     * Emits an {Approval} event indicating the updated allowance.
-     *
-     * Requirements:
-     *
-     * - `spender` cannot be the zero address.
+     * @dev See {IERC20Allowance-increaseAllowance}.
      */
-    function increaseAllowance(address spender, uint256 addedValue) public virtual returns (bool) {
+    function increaseAllowance(
+        address spender,
+        uint256 addedValue
+    ) public virtual override returns (bool)
+    {
         _approve(_msgSender(), spender, _allowances[_msgSender()][spender].add(addedValue));
         return true;
     }
 
     /**
-     * @dev Atomically decreases the allowance granted to `spender` by the caller.
-     *
-     * This is an alternative to {approve} that can be used as a mitigation for
-     * problems described in {IERC20-approve}.
-     *
-     * Emits an {Approval} event indicating the updated allowance.
-     *
-     * Requirements:
-     *
-     * - `spender` cannot be the zero address.
-     * - `spender` must have allowance for the caller of at least
-     * `subtractedValue`.
+     * @dev See {IERC20Allowance-decreaseAllowance}.
      */
-    function decreaseAllowance(address spender, uint256 subtractedValue) public virtual returns (bool) {
+    function decreaseAllowance(
+        address spender,
+        uint256 subtractedValue
+    ) public virtual override returns (bool)
+    {
         _approve(_msgSender(), spender, _allowances[_msgSender()][spender].sub(subtractedValue, "ERC20: decreased allowance below zero"));
         return true;
     }
+
+/////////////////////////////////////////// Internal Functions ///////////////////////////////////////
 
     /**
      * @dev Moves tokens `amount` from `sender` to `recipient`.
@@ -278,6 +264,8 @@ abstract contract ERC20 is Context, IERC165, IERC20, IERC20Detailed {
         _allowances[owner][spender] = amount;
         emit Approval(owner, spender, amount);
     }
+
+/////////////////////////////////////////// Hooks ///////////////////////////////////////
 
     /**
      * @dev Hook that is called before any transfer of tokens. This includes